Make it feel and look more like a dart library.
