+++ example.dart
@@ -1,6 +1,4 @@
-import 'package:poly_bool_dart/coordinate.dart';
-import 'package:poly_bool_dart/polybool.dart';
-import 'package:poly_bool_dart/types.dart';
+import 'package:polybool/polybool.dart';
 
 class LatLng {
   final double latitude;
