+++ lib/src/intersector.dart
