+++ lib/src/linked_list.dart
