+++ lib/src/segment_chainer.dart
