+++ lib/src/segment_fill.dart
