+++ lib/src/segment_selector.dart
