+++ pubspec.yaml
@@ -1,6 +1,6 @@
-name: poly_bool_dart
-description: Boolean operations on polygons (union, intersection, difference, xor) (this library is a port for flutter of polybooljs.
-version: 0.2.0
+name: polybool
+description: Boolean operations on polygons (union, intersection, difference, xor).
+version: 1.0.0
 homepage: https://github.com/mohammedX6/poly_bool_dart/tree/master
 
 environment:
